modify some route

+++ src/main/java/com/chatop/api/model/Rental.java
@@ -31,7 +31,7 @@ public class Rental {
     private String description;
     
     @Column(name="owner_id")
-    private int ownerId;
+    private Long ownerId;
     
     @Column(name="created_at")
     private Date createdAt;
